backend/pyproject.toml
[project]
name = "vidwiz"
version = "0.1.0"
description = ""
authors = [
    {name = "adhirajpandey",email = "pandey.adhiraj02@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<3.14"
dependencies = [
    "python-dotenv (>=1.1.1,<2.0.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "pydantic-settings (>=2.9.0,<3.0.0)",
    "fastapi (>=0.115.0,<1.0.0)",
    "uvicorn (>=0.30.0,<1.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "pytest (>=8.4.1,<9.0.0)",
    "boto3 (>=1.39.4,<2.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "youtube-transcript-api (>=1.2.2,<2.0.0)",
    "yt-dlp (>=2025.12.8,<2026.0.0)",
    "google-auth (>=2.20.0,<3.0.0)",
    "google-generativeai (>=0.8.6,<0.9.0)",
    "google-genai (>=1.60.0,<2.0.0)",
    "werkzeug (>=3.0.0,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
vidwiz = "src.main:run"

[tool.poetry]
package-mode = false

[tool.pytest.ini_options]
pythonpath = "."
testpaths = ["src/tests"]

[dependency-groups]
dev = [
    "pytest-cov (>=7.0.0,<8.0.0)"
]
